Module aws_sdk_datapipeline::error

Modules

internal_service_error

See InternalServiceError

invalid_request_exception

See InvalidRequestException

pipeline_deleted_exception

See PipelineDeletedException

pipeline_not_found_exception

See PipelineNotFoundException

task_not_found_exception

See TaskNotFoundException

Structs

ActivatePipelineError

AddTagsError

CreatePipelineError

DeactivatePipelineError

DeletePipelineError

DescribeObjectsError

DescribePipelinesError

EvaluateExpressionError

GetPipelineDefinitionError

InternalServiceError

An internal service error occurred.

InvalidRequestException

The request was not valid. Verify that your request was properly formatted, that the signature was generated with the correct credentials, and that you haven't exceeded any of the service limits for your account.

ListPipelinesError

PipelineDeletedException

The specified pipeline has been deleted.

PipelineNotFoundException

The specified pipeline was not found. Verify that you used the correct user and account identifiers.

PollForTaskError

PutPipelineDefinitionError

QueryObjectsError

RemoveTagsError

ReportTaskProgressError

ReportTaskRunnerHeartbeatError

SetStatusError

SetTaskStatusError

TaskNotFoundException

The specified task was not found.

ValidatePipelineDefinitionError

Enums

ActivatePipelineErrorKind

AddTagsErrorKind

CreatePipelineErrorKind

DeactivatePipelineErrorKind

DeletePipelineErrorKind

DescribeObjectsErrorKind

DescribePipelinesErrorKind

EvaluateExpressionErrorKind

GetPipelineDefinitionErrorKind

ListPipelinesErrorKind

PollForTaskErrorKind

PutPipelineDefinitionErrorKind

QueryObjectsErrorKind

RemoveTagsErrorKind

ReportTaskProgressErrorKind

ReportTaskRunnerHeartbeatErrorKind

SetStatusErrorKind

SetTaskStatusErrorKind

ValidatePipelineDefinitionErrorKind