Module aws_sdk_datapipeline::model

Modules

field

See Field

instance_identity

See InstanceIdentity

operator

See Operator

parameter_attribute

See ParameterAttribute

parameter_object

See ParameterObject

parameter_value

See ParameterValue

pipeline_description

See PipelineDescription

pipeline_id_name

See PipelineIdName

pipeline_object

See PipelineObject

query

See Query

selector

See Selector

tag

See Tag

task_object

See TaskObject

validation_error

See ValidationError

validation_warning

See ValidationWarning

Structs

Field

A key-value pair that describes a property of a pipeline object. The value is specified as either a string value (StringValue) or a reference to another object (RefValue) but not as both.

InstanceIdentity

Identity information for the EC2 instance that is hosting the task runner. You can get this value by calling a metadata URI from the EC2 instance.

Operator

Contains a logical operation for comparing the value of a field with a specified value.

ParameterAttribute

The attributes allowed or specified with a parameter object.

ParameterObject

Contains information about a parameter object.

ParameterValue

A value or list of parameter values.

PipelineDescription

Contains pipeline metadata.

PipelineIdName

Contains the name and identifier of a pipeline.

PipelineObject

Contains information about a pipeline object. This can be a logical, physical, or physical attempt pipeline object. The complete set of components of a pipeline defines the pipeline.

Query

Defines the query to run against an object.

Selector

A comparision that is used to determine whether a query should return this object.

Tag

Tags are key/value pairs defined by a user and associated with a pipeline to control access. AWS Data Pipeline allows you to associate ten tags per pipeline.

TaskObject

Contains information about a pipeline task that is assigned to a task runner.

ValidationError

Defines a validation error. Validation errors prevent pipeline activation. The set of validation errors that can be returned are defined by AWS Data Pipeline.

ValidationWarning

Defines a validation warning. Validation warnings do not prevent pipeline activation. The set of validation warnings that can be returned are defined by AWS Data Pipeline.

Enums

OperatorType

TaskStatus