Module aws_sdk_datapipeline::operation

Structs

ActivatePipeline

Validates the specified pipeline and starts processing pipeline tasks. If the pipeline does not pass validation,

AddTags

Adds or modifies tags for the specified pipeline.

CreatePipeline

Creates a new, empty pipeline. Use PutPipelineDefinition to populate the pipeline.

DeactivatePipeline

Deactivates the specified running pipeline. The pipeline is set to the DEACTIVATING

DeletePipeline

Deletes a pipeline, its pipeline definition, and its run history.

DescribeObjects

Gets the object definitions for a set of objects associated with the pipeline. Object definitions are composed of

DescribePipelines

Retrieves metadata about one or more pipelines. The information retrieved includes the name of the pipeline, the pipeline identifier,

EvaluateExpression

Task runners call EvaluateExpression to evaluate a string in the context of the specified object.

GetPipelineDefinition

Gets the definition of the specified pipeline. You can call GetPipelineDefinition to retrieve

ListPipelines

Lists the pipeline identifiers for all active pipelines that you have permission to access.

PollForTask

Task runners call PollForTask to receive a task to perform from AWS Data Pipeline. The task runner specifies which tasks it can perform

PutPipelineDefinition

Adds tasks, schedules, and preconditions to the specified pipeline. You can use PutPipelineDefinition to populate a new pipeline.

QueryObjects

Queries the specified pipeline for the names of objects that match the specified set of conditions.

RemoveTags

Removes existing tags from the specified pipeline.

ReportTaskProgress

Task runners call ReportTaskProgress when assigned a task to acknowledge that it has the task. If the web service does not

ReportTaskRunnerHeartbeat

Task runners call ReportTaskRunnerHeartbeat every 15 minutes to indicate that they are operational.

SetStatus

Requests that the status of the specified physical or logical pipeline objects be updated in the specified pipeline.

SetTaskStatus

Task runners call SetTaskStatus to notify AWS Data Pipeline that a task is completed and provide information about the final status.

ValidatePipelineDefinition

Validates the specified pipeline definition to ensure that it is well formed and can be run without error.